Add gas improvements

+++ contracts/lib/SkrybeLib.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.21;
 
 struct CollectionParams {
     string collectionId;
-    uint256 launchTimestamp;
-    uint256 whitelistLaunchTimestamp;
+    uint32 launchTimestamp;
+    uint32 whitelistLaunchTimestamp;
     uint256 price;
     uint256 whitelistPrice;
-    uint256 maxSupply;
-    uint256 maxPerTxn;
-    uint256 maxPerWhitelist;
-    uint256 maxPerWallet;
-    uint256 usesWhitelist;
+    uint32 maxSupply;
+    uint32 maxPerTxn;
+    uint32 maxPerWhitelist;
+    uint32 maxPerWallet;
+    uint8 usesWhitelist;
 }
